поправил тип

+++ protected/modules/account/migrations/m000099_000000_create_user.php
@@ -42,11 +42,11 @@ class m000099_000000_create_user extends Migration
 
             'status' => $this->smallInteger()->notNull()->defaultValue(User::STATUS_ACTIVE),
 
-            'created_at' => $this->integer()->notNull(),
+            'created_at' => $this->bigInteger()->notNull(),
             'created_ip' => $this->string(45),
 
-            'last_login_at' => $this->integer(),
-            'last_request_at' => $this->integer(),
+            'last_login_at' => $this->bigInteger(),
+            'last_request_at' => $this->bigInteger(),
 
             'session' => $this->string(),
         ], $this->tableOptions);
