+++ protected/modules/products/migrations/m000104_000001_create_page_products.php
@@ -29,7 +29,7 @@ class m000104_000001_create_page_products extends Migration
         $this->createTable($this->tableName, [
             'id' => $this->primaryKey(),
 
-            'published_at' => $this->integer(),
+            'published_at' => $this->bigInteger(),
             'status' => $this->smallInteger()->notNull()->defaultValue(Products::STATUS_DRAFT),
 
             'content_title' => $this->string()->notNull(),
